move output of restart FIP values into FIPContainer
