+++ opm/simulators/flow/GenericOutputBlackoilModule.cpp
@@ -697,53 +697,7 @@ assignToSolution(data::Solution& sol)
     }
 
     // Fluid in place
-    if (this->outputFipRestart_) {
-        using namespace std::string_literals;
-
-        using M = UnitSystem::measure;
-        using FIPEntry = std::tuple<std::string, M, Inplace::Phase>;
-
-        auto fipArrays = std::vector<FIPEntry> {};
-        if (this->outputFipRestart_.surface) {
-            fipArrays.insert(fipArrays.end(), {
-                    FIPEntry {"SFIPOIL"s, M::liquid_surface_volume, Inplace::Phase::OIL   },
-                    FIPEntry {"SFIPWAT"s, M::liquid_surface_volume, Inplace::Phase::WATER },
-                    FIPEntry {"SFIPGAS"s, M::gas_surface_volume,    Inplace::Phase::GAS   },
-                });
-        }
-
-        if (this->outputFipRestart_.reservoir) {
-            fipArrays.insert(fipArrays.end(), {
-                    FIPEntry {"RFIPOIL"s, M::volume, Inplace::Phase::OilResVolume   },
-                    FIPEntry {"RFIPWAT"s, M::volume, Inplace::Phase::WaterResVolume },
-                    FIPEntry {"RFIPGAS"s, M::volume, Inplace::Phase::GasResVolume   },
-                });
-        }
-
-        if (this->outputFipRestart_.noPrefix && !this->outputFipRestart_.surface) {
-            fipArrays.insert(fipArrays.end(), {
-                    FIPEntry { "FIPOIL"s, M::liquid_surface_volume, Inplace::Phase::OIL   },
-                    FIPEntry { "FIPWAT"s, M::liquid_surface_volume, Inplace::Phase::WATER },
-                    FIPEntry { "FIPGAS"s, M::gas_surface_volume,    Inplace::Phase::GAS   },
-                });
-        }
-
-        for (const auto& [mnemonic, unit, phase] : fipArrays) {
-            if (! this->fip_[phase].empty()) {
-                sol.insert(mnemonic, unit, std::move(this->fip_[phase]),
-                           data::TargetType::RESTART_SOLUTION);
-            }
-        }
-
-        for (const auto& phase : Inplace::mixingPhases()) {
-            if (! this->fip_[phase].empty()) {
-                sol.insert(Inplace::EclString(phase),
-                           UnitSystem::measure::volume,
-                           this->fip_[phase],
-                           data::TargetType::SUMMARY);
-            }
-        }
-    }
+    this->fipC_.outputRestart(sol);
 
     // Tracers
     if (! this->freeTracerConcentrations_.empty()) {
@@ -926,31 +880,10 @@ doAllocBuffers(const unsigned bufferSize,
     }
 
     // Fluid in place
-    {
-        using namespace std::string_literals;
-
-        const auto fipctrl = std::array {
-            std::pair { "FIP"s , &OutputFIPRestart::noPrefix  },
-            std::pair { "SFIP"s, &OutputFIPRestart::surface   },
-            std::pair { "RFIP"s, &OutputFIPRestart::reservoir },
-        };
-
-        this->outputFipRestart_.clearBits();
-        this->computeFip_ = false;
-
-        for (const auto& [mnemonic, kind] : fipctrl) {
-            if (auto fipPos = rstKeywords.find(mnemonic);
-                fipPos != rstKeywords.end())
-            {
-                fipPos->second = 0;
-                this->outputFipRestart_.*kind = true;
-            }
-        }
-
-        this->computeFip_ = this->fipC_.allocate(bufferSize,
-                                                 summaryConfig_,
-                                                 !substep);
-    }
+    this->computeFip_ = this->fipC_.allocate(bufferSize,
+                                             summaryConfig_,
+                                             !substep,
+                                             rstKeywords);
 
     const auto needAvgPress = !substep         ||
         !this->RPRNodes_.empty()               ||
